refactor(online-chat): Modularize chat client with SOLID principles and ES6 modules
