+++ online-chat-message/chat-client/src/index.html
@@ -61,6 +61,7 @@
       </div>
     </div>
 
-    <script src="app.js"></script>
+    <!-- JavaScriptをモジュールとして読み込む -->
+    <script type="module" src="app.js"></script>
   </body>
 </html>
