+++ online-chat-message/chat-client/src/models/User.js
@@ -1,0 +1,14 @@
+/**
+ * ユーザーモデル - ユーザー情報を管理
+ */
+class User {
+  constructor(username) {
+    this.username = username;
+  }
+
+  getUsername() {
+    return this.username;
+  }
+}
+
+export default User;
